fix: don't allow primitive types in handlers arguments (#39)
## Description

This PR aims to fix an issue in the generated TypeSpec that occasionally
generated handlers whose arguments were primitive types (e.g. `date`).
However, the handler argument needs to be a complex object (a TSP
Model), so we can parse it and transform it into JSON.

**FIX**

- Provided an extra line to the prompt, to make sure this occurs way
less often;
- Through some TSP decorator reflection, throw a compilation error if
this occurs;


## Test Plan

Tested this fix by forcing the generation of these "bugged" handlers and
looked at the Traces to understand how the bot would react in the next
iterations.

Example tracing:
https://cloud.langfuse.com/project/cm6j91sap009ux891llzvdjvi/traces/8d70aaf4-d656-4aaa-b09e-332a03bec9a9?timestamp=2025-02-25T17%3A20%3A11.295Z


![image](https://github.com/user-attachments/assets/ab7d8ea3-45d9-4aff-8591-a98b092220f3)