fix: remove unused vars with tool (#63)
## Description

This PR aims to remove any unused generated variables with a tool
instead of relying on LLM to do another generation.

## Test Plan

I tested this by manually adding an unused variable to the test template
and generating a bot. Because we remove the variables before running
lint, we never get to see errors related to that in the `stderr` output.