fix: migrations that require interaction with CLI (#286)
## Description

This PR aims to fix an issue for DB migrations that require some sort of
interaction in the CLI. The fix is kinda hacky, as it just uses the
first option everytime, but should do for the time being.

The "optimal fix" would need to envolve the agent to actually select the
proper CLI option and generate migrations file, according to what was
actually done. cc @igrekun